Fix a persistent bug with non-interactive trails at metadata zooms.
The repro was zoom into a trail, go to it's detail level containing
boundary, refresh, go back, zoom out to metadata zoom.

The root cause was that the trail was loaded from a metadata cell at
zoom X and a detail cell at zoom Y. So when going back the detail cell
got unloaded, causing the trail to be deleted from both detail and
metadata, while the metadata cell was kept loaded. Then when we zoomed
out, the metadata cell was already considered loaded so we never
reloaded the trail.