Allow hovering on the elevation graph
I am not very happy with this solution. I am recording the lat/lngs of
each sample point server-side, so it leads to duplicate data fetches
since practically we only care about lat-lngs when hovering on a path on
the map. The alternative is duplicating the elevation calculation
routine in JS, which does make some sense but is annoying.

Will think about it.