Replace the ID trail URL with its readable name.
This is all gross. I wanted to be clever and stick with bigints on the
client as much as possible, but this just means a bunch of joining on
the database and total confusion when fetching results from the server.
Plus we end up making space for the readable ID even if we don't
populate it. I think it's better to just eat the cost.